refactor(cli): split boxdraw.go into frame, sections, and advisory files
boxdraw.go (346 LOC) now contains only the box-drawing primitives, wrap
helpers, title/verdict helpers, and the two entry-point orchestrators.

boxdraw_sections.go (664 LOC) holds all ten writeBox* section renderers
plus packageEcoName — the audit-entry display logic layer.

boxdraw_advisory.go (126 LOC) holds the four advisory formatting helpers
(sortedAdvisoryBlock, formatAdvisoryLines, advisoryCountText,
formatTransitiveAdvisoryLines) that are co-used by Releases rendering.

No signature changes; boxdraw_test.go compiles and passes unchanged.
truncateDescription remains in input_helpers.go (phase 1).
